EventPlatformNAG/requirements.txt
antlr4-python3-runtime==4.13.0
antlr4-tools==0.2
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
cattrs==25.2.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.7
dnspython==2.8.0
email-validator==1.3.0
Flask==2.2.2
Flask-Cors==3.0.4
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.0.2
Flask-User==1.0.2.2
Flask-WTF==1.2.2
forbiddenfruit==0.1.4
greenlet==3.2.4
idna==3.10
iniconfig==2.1.0
install-jdk==1.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
lsprotocol==2025.0.0
MarkupSafe==3.0.2
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
pycparser==2.22
pygls==2.0.0a6
pytest==7.2.0
six==1.17.0
SQLAlchemy==2.0.43
typing_extensions==4.15.0
Werkzeug==2.2.2
WTForms==3.2.1